requirements.txt
astroid==2.5.2
click==7.1.2
Flask==1.1.2
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.7.4
pyserial==3.5
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
